+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 6.12.0
+
+* added funnel
+* updated referrer blacklist
+* updated User-Agent blacklist
+* updated dependencies
+
 ## 6.11.1
 
 * added `not contains` filter option